# api-specification.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Graduation Project API Specification",
    "version": "1.0.0",
    "description": "API specification for Graduation Project Backend - Built with Spring Boot 3.5.7 and Java 21, implementing Hexagonal Architecture (Ports & Adapters Pattern)",
    "contact": {
      "name": "manhuynh",
      "email": "22120201@student.hcmus.edu.vn"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Local development server"
    },
    {
      "url": "https://api.graduationproject.com",
      "description": "Production server"
    }
  ],
...
  "paths": {
...
    "/api/exams": {
      "post": {
        "tags": [
          "Exams"
        ],
        "summary": "Create a new exam",
        "description": "Create a new exam for a class. Requires TEACHER role.",
        "operationId": "createExam",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateExamRequestDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Exam created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden - requires TEACHER role"
          }
        }
      }
    },
    "/api/exams/{examId}": {
      "get": {
        "tags": [
          "Exams"
        ],
        "summary": "Get exam detail",
        "description": "Get detail of a specific published exam. Requires STUDENT role and enrollment in the exam's class.",
        "operationId": "getExamDetail",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "examId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Exam retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden - requires STUDENT role"
          },
          "404": {
            "description": "Exam not found"
          }
        }
      }
    }
  },
  "components": {
...
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT token obtained from login endpoint. Include in Authorization header as: Bearer {token}"
      }
    }
  },
  "security": []
}